pyproject.toml
[project]
name = "age-by-face"
version = "0.1.0"
description = "Service for determining age from a face photo"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc>=3.64.2",
    "dvc-s3>=3.2.2",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "lightning>=2.6.0",
    "mlflow>=3.7.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]

[dependency-groups]
dev = [
    "fire>=0.7.1",
    "pre-commit>=4.5.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.9",
]

[project.scripts]
age-by-face = "age_by_face.commands:main"

[tool.ruff]
line-length = 100
format = { docstring-code-format = true }

[tool.ruff.lint]
select = [
    # isort
    "I",
    # pycodestyle
    "E", "W",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # flake8-builtins
    "A",
    # flake8-return
    "RET",
    # flake8-use-pathlib
    "PTH",
    # flynt
    "FLY",
    # mccabe
    "C90",
    # NumPy-specific rules
    "NPY",
    # pep8-naming
    "N",
    # Ruff-specific rules
    "RUF",
    # Pylint - very strict, may be use when have enough time
    "PL",
    # refurb - also may be overwhelming
    "FURB",
]

pycodestyle = { max-doc-length = 105 }

# [tool.ruff.lint.per-file-ignores]
# "__init__.py" = ["E402"]
# "**/{tests,docs,tools}/*" = ["E402"]
